backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Configuration management
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-decouple==3.8
PyYAML==6.0.1

# Database and caching
psycopg[binary]==3.2.9
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
redis[hiredis]==5.0.1

# Time series database (TimescaleDB is PostgreSQL extension)
# psycopg2-binary==2.9.9 # Replaced with psycopg

# Trading and financial data
numpy==1.23.5
pandas==2.1.4
aiohttp==3.9.1
yfinance==0.2.18
kiteconnect==5.0.1

# Additional dependencies for NSE data management
beautifulsoup4==4.12.2  # For web scraping if needed
lxml==4.9.3  # XML/HTML parsing

# Technical Analysis - Modern alternatives (No C dependencies!)
pandas-ta==0.3.14b0

# Advanced analytics and backtesting
vectorbt==0.28.1
backtrader==1.9.78.123

# WebSocket support
websockets==12.0
python-socketio==5.10.0

# Real-time data processing
pandas==2.1.4
numpy==1.23.5
scipy==1.11.4

# HTTP clients
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Async support
asyncio==3.4.3
aioredis==2.0.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and logging
structlog==23.2.0
sentry-sdk==1.38.0
